added the useWindowWidth custom hook
