+++ src/components/hooks/useWindowWidth.js
@@ -1,0 +1,19 @@
+import { useState, useEffect, useCallback } from 'react';
+
+export function useWindowWidth() {
+    const [innerWidth, setInnerWidth] = useState(window.innerWidth);
+
+    const handleResize = useCallback(() => {
+        setInnerWidth(window.innerWidth);
+    }, []);
+
+    useEffect(() => {
+        window.addEventListener('resize', handleResize);
+        setInnerWidth(window.innerWidth);
+        return () => {
+            window.removeEventListener('resize', handleResize);
+        };
+    }, [handleResize]);
+
+    return { innerWidth };
+}
